Update README.md
Updated python.README to better clarify where data-structures or code-challenge files should go. The original directions were to make a folder for data_structures or code_challenges and create a file within those. data_structures and code_challenges folders already have the files pre-made within them. So I updated the README to better clarify that and make it a little easier to follow for future students